+++ pyfortest/test_pyfor.py
@@ -232,6 +232,9 @@ class RasterTestCase(unittest.TestCase):
         self.test_raster.plot()
         self.test_raster.plot(return_plot=True)
 
+    def test_local_maxima(self):
+        self.test_raster.local_maxima()
+
     def test_write_with_crs(self):
         self.test_raster.write("./thing.tif")
         os.remove("./thing.tif")
@@ -239,6 +242,14 @@ class RasterTestCase(unittest.TestCase):
     def test_write_without_crs(self):
         self.test_raster.crs = None
         self.test_raster.write("./thing.tif")
+
+class DetectedTopsTestCase(unittest.TestCase):
+    def setUp(self):
+        self.test_detect = cloud.Cloud(test_las).chm(1, interp_method='nearest').local_maxima()
+
+    def test_plot(self):
+        self.test_detect.plot()
+
 
 def test_func(las_path):
     cloud.Cloud(las_path)